requirements.txt
#
# Requirements for wger for production
#

Django>=1.9,<1.10
django-recaptcha==1.3.1
social-auth-app-django==3.1.0
reportlab==3.3.0
django_mobile
django-formtools>=1.0,<1.1
bleach
python-mimeparse
pillow
easy-thumbnails
django_compressor
icalendar
django-bootstrap-breadcrumbs
sphinx
django-sortedm2m
django-bower
invoke==1.2.0
requests
python-dateutil
requests-oauthlib
fitbit==0.3.1
# REST API
djangorestframework>=3.2,<3.3
django-filter==1.1.0
django-tastypie
django-cors-headers==2.4.1 

# Python3 compatibility
six
psycopg2
psycopg2_binary
whitenoise==4.1.2
gunicorn==19.9.0